requirements.txt
pydub-0.25.1
matplotlib-3.6.3
librosa-0.9.2
sounddevice-0.4.5
pyaudio-0.2.13
py-structs-1.1.0
opencv-python-4.7.0.68